+++ src/app.h
@@ -15,8 +15,6 @@
 #include "hm/hmSystem.h"
 #include "hm/hmRadio.h"
 #include "hms/hmsRadio.h"
-//#include "hm/hmPayload.h"
-//#include "hm/miPayload.h"
 #include "publisher/pubMqtt.h"
 #include "publisher/pubSerial.h"
 #include "utils/crc.h"
@@ -41,8 +39,6 @@
 #define ACOS(x) (degrees(acos(x)))
 
 typedef HmSystem<MAX_NUM_INVERTERS> HmSystemType;
-//typedef HmPayload<HmSystemType> PayloadType;
-//typedef MiPayload<HmSystemType> MiPayloadType;
 #ifdef ESP32
 typedef CmtRadio<esp32_3wSpi> CmtRadioType;
 #endif
@@ -169,15 +165,6 @@ class app : public IApp, public ah::Scheduler {
 
         void setMqttPowerLimitAck(Inverter<> *iv) {
             mMqtt.setPowerLimitAck(iv);
-        }
-
-        void ivSendHighPrio(Inverter<> *iv) {
-            if(mIVCommunicationOn) { // only send commands if communication is enabled
-                //if (iv->ivGen == IV_MI)
-                //    mMiPayload.ivSendHighPrio(iv);
-                //else
-                //    mPayload.ivSendHighPrio(iv);
-            }
         }
 
         bool getMqttIsConnected() {